+++ library/src/main/java/org/dbtools/android/room/CloseableDatabaseWrapperRepository.kt
@@ -179,28 +179,4 @@ abstract class CloseableDatabaseWrapperRepository<out T: RoomDatabase>(protected
         keysCopy.addAll(databaseList.keys)
         keysCopy.forEach { key -> closeDatabase(key, deleteFiles) }
     }
-
-    fun dropView(database: SupportSQLiteDatabase, viewName: String) {
-        DatabaseUtil.dropView(database, viewName)
-    }
-
-    fun dropAllViews(database: SupportSQLiteDatabase, views: List<DatabaseViewQuery>) {
-        DatabaseUtil.dropAllViews(database, views)
-    }
-
-    fun createView(database: SupportSQLiteDatabase, viewName: String, viewQuery: String) {
-        DatabaseUtil.createView(database, viewName, viewQuery)
-    }
-
-    fun createAllViews(database: SupportSQLiteDatabase, views: List<DatabaseViewQuery>) {
-        DatabaseUtil.createAllViews(database, views)
-    }
-
-    fun recreateView(database: SupportSQLiteDatabase, viewName: String, viewQuery: String) {
-        DatabaseUtil.recreateView(database, viewName, viewQuery)
-    }
-
-    fun recreateAllViews(database: SupportSQLiteDatabase, views: List<DatabaseViewQuery>) {
-        DatabaseUtil.recreateAllViews(database, views)
-    }
 }